+++ example/lib/main.dart
@@ -93,6 +93,10 @@ class _MyAppState extends State<MyApp> {
           // Installed from iOS app store
           source = "App Store";
           break;
+        case Source.IS_PENDING_APP_STORE_REVIEW:
+          // Installed version is not available on the store yet
+          source = "Pending Release";
+          break;
         case Source.IS_INSTALLED_FROM_TEST_FLIGHT:
           // Installed from Test Flight
           source = "Test Flight";